fix log format and parameter mismatched bug (check by printf)
git-svn-id: http://fastdfs.googlecode.com/svn@60 0d4c4929-bc51-0410-a161-b7c81ae5c018

+++ trunk/common/shared_func.c
@@ -974,7 +974,7 @@ int set_rlimit(int resource, const rlim_t value)
 		logError("file: "__FILE__", line: %d, " \
 			"call setrlimit fail, resource=%d, value=%d, " \
 			"errno: %d, error info: %s", \
-			__LINE__, resource, value, \
+			__LINE__, resource, (int)value, \
 			errno, strerror(errno));
 		return errno != 0 ? errno : EPERM;
 	}
